pyproject.toml
[tool.poetry]
name = "users_microservice"
version = "0.1.2"
description = "Users microservice for BookBNB."
authors = ["75.52 - 2020C2 - G5 <FIUBA email>"]

[tool.poetry.dependencies]
python = "^3.7"
sqlalchemy = "^1.3.19"
argparse = "^1.4.0"
importlib-metadata = {version = "^1.0", python = "<3.8"}
pytest = {version = "^6.1.0", optional = true}
pytest-cov = {version = "^2.10.1", optional = true}
pytest-xdist = {version = "^2.1.0", optional = true}
youconfigme = "^0.6.14"
flask = "^1.1.2"
flask-restx = "^0.2.0"
flask-migrate = "^2.5.3"
flask-sqlalchemy = "^2.4.4"
flask_script = "^2.0.6"
sqlalchemy_utils = "^0.36.8"
psycopg2-binary = "^2.8.6"
pylint-flask = "^0.6"
pylint-flask-sqlalchemy = "^0.2.0"
SQLAlchemy-Utils = "^0.36.8"
Flask-Cors = "^3.0.9"
Flask-Bcrypt = "^0.7.1"
PyJWT = "^1.7.1"

[tool.poetry.dev-dependencies]
nox = "^2020.8.22"
pylint = "^2.6.0"
mypy = "^0.782"
flake8 = "^3.8.3"
black = "19.10b0"
pre-commit = "^2.7.1"
bandit = "^1.6.2"
bump2version = "^1.0.1"

[tool.poetry.extras]
testing = ["pytest", "pytest-cov", "pytest-xdist"]

[tool.black]
line-length = 88
skip-string-normalization = true

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta:__legacy__"

